requirements.txt
django-crispy-forms
crispy-bootstrap5
django-tinymce